ASG6: Moved files to correct folder; fixed namespace;
