--- Editor/Resources/EditorWindow/GameLiftPlugin.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: e2c810574e47434dada447ccfcddc894
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences:
-  - m_ViewDataDictionary: {instanceID: 0}
-  - icon: {fileID: 2800000, guid: 31e3daa037125ad4696a676394e5c633, type: 3}
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
